Dispatcher Contract　ファイル分離
